Move DetermineUserJobFactory into creds
